tests/wpt/web-platform-tests/secure-payment-confirmation/constructor.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<meta name="timeout" content="long">
<title>Test for the 'secure-payment-confirmation' payment method constructor</title>
<link rel="help" href="https://w3c.github.io/secure-payment-confirmation/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

const details = {total:
    {label: 'Total', amount: {value: '0.01', currency: 'USD'}}};

test(() => {
  new PaymentRequest([{
    supportedMethods: 'secure-payment-confirmation',
    data: {
       // All valid parameters.
       credentialIds: [Uint8Array.from('x', c => c.charCodeAt(0))],
       challenge: Uint8Array.from('x', c => c.charCodeAt(0)),
       payeeOrigin: window.location.origin,
       timeout: 60000,
       instrument: {
        displayName: 'X',
        icon: 'https://example.test/icon.png',
       },
    },
  }], details);
}, 'Valid payment method data does not throw exceptions.');

test(() => {
  new PaymentRequest([{
    supportedMethods: 'secure-payment-confirmation',
    data: {
       credentialIds: [Uint8Array.from('x', c => c.charCodeAt(0))],
       challenge: Uint8Array.from('x', c => c.charCodeAt(0)),
       payeeOrigin: window.location.origin,
       // Omitted timeout field.
       instrument: {
         displayName: 'X',
         icon: 'https://example.test/icon.png',
       },
    },
  }], details);
}, 'The timeout field is optional.');

test(() => {
  assert_throws_js(RangeError, () => {
    new PaymentRequest([{
      supportedMethods: 'secure-payment-confirmation',
      data: {
        credentialIds: [Uint8Array.from('x', c => c.charCodeAt(0))],
        challenge: Uint8Array.from('x', c => c.charCodeAt(0)),
        payeeOrigin: window.location.origin,
        timeout: 60000,
        instrument: {
          displayName: 'X',
          icon: 'https://example.test/icon.png',
        },
      },
    }, {supportedMethods: 'basic-card'}], details);
  });
}, 'Extra payment method not allowed afterward.');

test(() => {
  assert_throws_js(RangeError, () => {
    new PaymentRequest([{supportedMethods: 'basic-card'}, {
      supportedMethods: 'secure-payment-confirmation',
      data: {
        credentialIds: [Uint8Array.from('x', c => c.charCodeAt(0))],
        challenge: Uint8Array.from('x', c => c.charCodeAt(0)),
        payeeOrigin: window.location.origin,
        timeout: 60000,
        instrument: {
          displayName: 'X',
          icon: 'https://example.test/icon.png',
        },
      },
    }], details);
  });
}, 'Extra payment method not allowed beforehand.');

test(() => {
  assert_throws_js(TypeError, () => {
    new PaymentRequest([{
      supportedMethods: 'secure-payment-confirmation',
      data: {
        // Omitted credentialIds field.
        challenge: Uint8Array.from('x', c => c.charCodeAt(0)),
        payeeOrigin: window.location.origin,
        timeout: 60000,
        instrument: {
          displayName: 'X',
          icon: 'https://example.test/icon.png',
        },
      },
    }], details);
  });
}, 'The credentialIds field is required.');

test(() => {
  assert_throws_js(TypeError, () => {
    new PaymentRequest([{
      supportedMethods: 'secure-payment-confirmation',
      data: {
        credentialIds: [Uint8Array.from('x', c => c.charCodeAt(0))],
        payeeOrigin: window.location.origin,
        // Omitted challenge field.
        timeout: 60000,
        instrument: {
          displayName: 'X',
          icon: 'https://example.test/icon.png',
        },
      },
    }], details);
  });
}, 'The challenge field is required.');

test(() => {
  assert_throws_js(RangeError, () => {
    new PaymentRequest([{
      supportedMethods: 'secure-payment-confirmation',
      data: {
        relyingPartyId: 'relying-party.example',
        // Empty credentialIds field.
        credentialIds: [],
        challenge: Uint8Array.from('x', c => c.charCodeAt(0)),
        payeeOrigin: window.location.origin,
        timeout: 60000,
        instrument: {
          displayName: 'X',
          icon: 'https://example.test/icon.png',
        },
      },
    }], details);
  });
}, 'Empty credentialIds field throws exception.');

test(() => {
  assert_throws_js(TypeError, () => {
    new PaymentRequest([{
      supportedMethods: 'secure-payment-confirmation',
      data: {
        credentialIds: [Uint8Array.from('x', c => c.charCodeAt(0))],
        // Null challenge fields.
        challenge: null,
        payeeOrigin: window.location.origin,
        timeout: 60000,
        instrument: {
          displayName: 'X',
          icon: 'https://example.test/icon.png',
        },
      },
    }], details);
  });
}, 'Null challenge field throws exception.');

test(() => {
  assert_throws_js(TypeError, () => {
    new PaymentRequest([{
      supportedMethods: 'secure-payment-confirmation',
      data: {
        credentialIds: [Uint8Array.from('x', c => c.charCodeAt(0))],
        payeeOrigin: window.location.origin,
        // Empty challenge fields.
        challenge: [],
        timeout: 60000,
        instrument: {
          displayName: 'X',
          icon: 'https://example.test/icon.png',
        },
      },
    }], details);
  });
}, 'Empty challenge field throws exception.');

test(() => {
  assert_throws_js(RangeError, () => {
    new PaymentRequest([{
      supportedMethods: 'secure-payment-confirmation',
      data: {
        credentialIds: [Uint8Array.from('x', c => c.charCodeAt(0))],
        challenge: Uint8Array.from('x', c => c.charCodeAt(0)),
        payeeOrigin: window.location.origin,
        // Timeout longer than 1 hour.
        timeout: 1000 * 60 * 60 + 1,
        instrument: {
          displayName: 'X',
          icon: 'https://example.test/icon.png',
        },
      },
    }], details);
  });
}, 'Timeout longer than 1 hour throws exception.');

test(() => {
  assert_throws_js(TypeError, () => {
    new PaymentRequest([{
      supportedMethods: 'secure-payment-confirmation',
      data: {
        // Large credentialIds value.
        credentialIds: [Uint8Array.from(
            'x'.repeat(1024 * 1024), c => c.charCodeAt(0))],
        challenge: Uint8Array.from('x', c => c.charCodeAt(0)),
        payeeOrigin: window.location.origin,
        timeout: 60000,
        instrument: {
          displayName: 'X',
          icon: 'https://example.test/icon.png',
        },
      },
    }], details);
  });
}, 'Large credentialIds value throws exception.');

test(() => {
  assert_throws_js(TypeError, () => {
    new PaymentRequest([{
      supportedMethods: 'secure-payment-confirmation',
      data: {
        credentialIds: [Uint8Array.from('x', c => c.charCodeAt(0))],
        // Large challenge value.
        challenge: Uint8Array.from('x'.repeat(1024 * 1024), c => c.charCodeAt(0)),
        payeeOrigin: window.location.origin,
        timeout: 60000,
        instrument: {
          displayName: 'X',
          icon: 'https://example.test/icon.png',
        },
      },
    }], details);
  });
}, 'Large challenge value throws exception.');

test(() => {
  assert_throws_js(TypeError, () => {
    new PaymentRequest([{
      supportedMethods: 'secure-payment-confirmation',
      data: {
        credentialIds: [Uint8Array.from('x', c => c.charCodeAt(0))],
        challenge: Uint8Array.from('x', c => c.charCodeAt(0)),
        payeeOrigin: window.location.origin,
        timeout: 60000,
        // Omitted instrument field.
      },
    }], details);
  });
}, 'Instrument field is required.');

test(() => {
  assert_throws_js(TypeError, () => {
    new PaymentRequest([{
      supportedMethods: 'secure-payment-confirmation',
      data: {
        credentialIds: [Uint8Array.from('x', c => c.charCodeAt(0))],
        challenge: Uint8Array.from('x', c => c.charCodeAt(0)),
        payeeOrigin: window.location.origin,
        timeout: 60000,
        instrument: {
          // Ommitted instrument display name.
          icon: 'https://example.test/icon.png',
        },
      },
    }], details);
  });
}, 'Instrument display name is required.');

test(() => {
  assert_throws_js(TypeError, () => {
    new PaymentRequest([{
      supportedMethods: 'secure-payment-confirmation',
      data: {
        credentialIds: [Uint8Array.from('x', c => c.charCodeAt(0))],
        challenge: Uint8Array.from('x', c => c.charCodeAt(0)),
        payeeOrigin: window.location.origin,
        timeout: 60000,
        instrument: {
          displayName: 'X',
          // Ommitted instrument icon.
        },
      },
    }], details);
  });
}, 'Instrument icon is required.');

test(() => {
  assert_throws_js(TypeError, () => {
    new PaymentRequest([{
      supportedMethods: 'secure-payment-confirmation',
      data: {
        // Omitted payee origin.
        credentialIds: [Uint8Array.from('x', c => c.charCodeAt(0))],
        challenge: Uint8Array.from('x', c => c.charCodeAt(0)),
        timeout: 60000,
        instrument: {
          displayName: 'X',
          icon: 'https://example.test/icon.png',
        },
      },
    }], details);
  });
}, 'Payee origin is required.');

</script>
